user info component to standalone
