+++ lib/process-services/src/lib/process-user-info/process-user-info.component.ts
@@ -15,16 +15,23 @@
  * limitations under the License.
  */
 
-import { UserInfoMode } from '@alfresco/adf-core';
+import { FullNamePipe, InitialUsernamePipe, UserInfoMode } from '@alfresco/adf-core';
 import { EcmUserModel, PeopleContentService } from '@alfresco/adf-content-services';
 import { Component, Input, OnDestroy, ViewChild, ViewEncapsulation } from '@angular/core';
-import { MatMenuTrigger, MenuPositionX, MenuPositionY } from '@angular/material/menu';
+import { MatMenuModule, MatMenuTrigger, MenuPositionX, MenuPositionY } from '@angular/material/menu';
 import { Subject } from 'rxjs';
 import { PeopleProcessService } from '../common/services/people-process.service';
 import { UserRepresentation } from '@alfresco/js-api';
+import { CommonModule } from '@angular/common';
+import { MatButtonModule } from '@angular/material/button';
+import { MatTabsModule } from '@angular/material/tabs';
+import { TranslateModule } from '@ngx-translate/core';
+import { MatCardModule } from '@angular/material/card';
 
 @Component({
     selector: 'adf-process-user-info',
+    standalone: true,
+    imports: [CommonModule, FullNamePipe, MatButtonModule, MatMenuModule, InitialUsernamePipe, MatTabsModule, TranslateModule, MatCardModule],
     templateUrl: './process-user-info.component.html',
     styleUrls: ['./process-user-info.component.scss'],
     encapsulation: ViewEncapsulation.None
